// front-end/src/types.ts
export type TokenSymbol = typeof tokenSymbol[keyof typeof tokenSymbol];

export type ChainPropType = {
  [index: string]: ChainProps;
};

export interface ChainProps {
  'blockTime': number;
  'logo'?: any;
  'ss58Format': number;
  'tokenDecimals': number;
  'tokenSymbol': TokenSymbol;
  'chainId': number;
  'rpcEndpoint': string;
  'category': string;
}

export type ChainLinksType = {
  [index: string]: ChainLinks;
};

export interface ChainLinks {
  blockExplorer: string;
  homepage: string;
  github: string;
  discord: string;
  twitter: string;
  telegram: string;
  youtube: string;
  reddit: string;
}